finished user seeder with faker

+++ app/Http/Controllers/PostsController.php
@@ -6,6 +6,8 @@ use Illuminate\Http\Request;
 
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
+use App\Models\Post;
+use Log;
 
 class PostsController extends Controller
 {
@@ -16,7 +18,7 @@ class PostsController extends Controller
      */
     public function index()
     {
-        $posts = \App\Models\Post::paginate(4);
+        $posts = Post::paginate(4);
         $data['posts']= $posts;
         return view('posts.index',$data);
 
@@ -40,13 +42,12 @@ class PostsController extends Controller
      */
     public function store(Request $request)
     {
-
-        $this->validate($request, \App\Models\Post::$rules);
+        $this->validate($request, Post::$rules);
 
         $title = $request->input('title');
         $content = $request->input('content');
         $url = $request->input('url');
-        $post = new \App\Models\Post();
+        $post = new Post();
         $post->title = $title;
         $post->content = $content;
         $post->url = $url;
@@ -54,6 +55,8 @@ class PostsController extends Controller
         $post->save();
 
         $request->session()->flash('successMessage', 'Post created');
+
+        Log::info("$title, $content, $url");
 
         return redirect()->action('PostsController@index');
     }
@@ -66,7 +69,12 @@ class PostsController extends Controller
      */
     public function show($id)
     {
-        $post = \App\Models\Post::find($id);
+        $post = Post::find($id);
+
+        if(!$post){
+            abort(404);
+        }
+
         $data['post'] = $post;
         return view('posts.show',$data);
     }
@@ -79,7 +87,12 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
-        $post = \App\Models\Post::find($id);
+        $post = Post::find($id);
+
+        if(!$post){
+            abort(404);
+        }
+
         $data['post'] = $post;
         return view('posts.edit',$data);
     }
@@ -93,11 +106,14 @@ class PostsController extends Controller
      */
     public function update(Request $request, $id)
     {
-        
+        $this->validate($request, Post::$rules);
 
-        $this->validate($request, \App\Models\Post::$rules);
+        $post = Post::find($id);
 
-        $post = \App\Models\Post::find($id);
+        if(!$post){
+            abort(404);
+        }
+
         $post->title = $request->title;
         $post->content =$request->content;
         $post->url =$request->url;
@@ -117,10 +133,13 @@ class PostsController extends Controller
      */
     public function destroy(Request $request, $id)
     {
+        $post = Post::find($id);
 
-        $post = \App\Models\Post::find($id);
+        if(!$post){
+            abort(404);
+        }
+
         $post->delete();
-
         $request->session()->flash('successMessage', 'Post deleted');
         return redirect()->action('PostsController@index');
     }
